RN_BOMB Disarming
This code will arm the bomb when the player attempts to delete the rune.
This will cause it to count down normally and a message appears stating
"Attempting to disarm 3... 2... 1... DISARMED!" if successful, however 1
out of 10 times it will detonate saying "Attempting to disarm 3... 2...
1... oops!".  This only happens when the user attempts to delete the
rune.  The idea is to scare players away since they will think you're
going to explode but also add a risk in removing the rune.  Note: When
looking to adjust func.qc I could not find it in this source code, is it
missing?